Removed the .load class helper
I don't see why this is needed any longer now that the init can handle it.

Ultimately, I am unsure what the best approach would be to pickling the classes that provide a memory mapped wrapper around the PINK binary files. I feel like serialising the entire binary file is not the right thing to do. So, for the moment I cam keeping the functionality away until I know what is best.