rearranged functions and __getstate__ for the _local
Using a class attribute in the Annotation class to keep track of the user defined labels created through the Annotator.interactive_annotation() process. But, this can not be pickled because it is not an instance attribute. The hacky len check helps to ensure that the object remain picklable many times (say as data is serialised for multiprocessing)