save at emd of loop and more 'repair' functionality
save and the end of the driving loop within interactive annotation, not at the start. This had the little nasty side effect of not saving the very last annotation. And since that annotation was not save, subsequent load functions could not load any `labels`, even if I forced a new `Annotation` object to be generated.