removed some typing information to help mypy and avoid isinstance checks
When making the setstate and getstate methods for pickle I changed the return type from the unique labels function. The general consensus is that using the Union[] to denote multiple possible returns types is not best practise and can be a little misleading, as this technically means a type combining a collection of other types.